refactor(video-player): consolidate iOS/macOS Swift sources into shared darwin target
- Move all Swift source files from ios/Classes/ and macos/Classes/ into
  darwin/divine_video_player/Sources/divine_video_player/
- Replace ios/ and macos/ podspecs with a single darwin/ podspec that
  targets both platforms (iOS 16.0, macOS 13.0)
- Add Package.swift for Swift Package Manager support
- Set sharedDarwinSource: true for both ios and macos in pubspec.yaml
- Update apple_threading_contract_test.dart to read from the unified
  darwin source path (removes the per-platform loop)
