+++ mobile/ios/NotificationServiceExtension/NotificationServiceExtension.entitlements
@@ -2,9 +2,7 @@
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
 <dict>
-    <key>com.apple.security.application-groups</key>
-    <array>
-        <string>group.co.openvine.app</string>
-    </array>
+	<key>com.apple.security.application-groups</key>
+	<array/>
 </dict>
 </plist>
