--- mobile/packages/divine_video_player/macos/divine_video_player.podspec
@@ -1,15 +1,0 @@
-Pod::Spec.new do |s|
-  s.name             = 'divine_video_player'
-  s.version          = '0.1.0'
-  s.summary          = 'Multi-clip seamless video player for macOS.'
-  s.description      = 'Multi-clip video player using AVFoundation.'
-  s.homepage         = 'https://github.com/divinevideo/divine-mobile'
-  s.license          = { :type => 'MIT' }
-  s.author           = { 'Divine Video' => 'dev@divinevideo.co' }
-  s.source           = { :http => 'https://github.com/divinevideo/divine-mobile' }
-  s.source_files     = 'Classes/**/*'
-  s.dependency       'FlutterMacOS'
-  s.platform         = :osx, '13.0'
-  s.swift_version    = '5.9'
-  s.pod_target_xcconfig = { 'DEFINES_MODULE' => 'YES' }
-end
