+++ mobile/packages/divine_video_player/test/src/apple_threading_contract_test.dart
@@ -4,94 +4,90 @@ import 'package:flutter_test/flutter_test.dart';
 
 void main() {
   group('Apple native player threading contract', () {
-    for (final platform in ['ios', 'macos']) {
-      test('$platform sends Flutter event updates on the main thread', () {
-        final source = _appleSourceFile(platform).readAsStringSync();
+    test('sends Flutter event updates on the main thread', () {
+      final source = _appleSourceFile().readAsStringSync();
 
-        expect(
-          source,
-          contains('Thread.isMainThread'),
-          reason:
-              'FlutterEventSink must not be called from AVFoundation/KVO '
-              'callback queues.',
-        );
-        expect(source, contains('DispatchQueue.main.async'));
-        expect(source, contains('sendStateUpdateOnMain'));
-      });
-
-      test('$platform uses native queue-based looping', () {
-        final source = _appleSourceFile(platform).readAsStringSync();
-
-        expect(
-          source,
-          contains('AVQueuePlayer'),
-          reason:
-              'Apple loop playback should use a queue player so the next '
-              'iteration can be prepared before the current one ends.',
-        );
-        expect(
-          source,
-          contains('AVPlayerLooper'),
-          reason:
-              'Manual AVPlayerItemDidPlayToEndTime seek-and-replay loops '
-              'can create an end-of-item playback gap.',
-        );
-        expect(
-          source,
-          isNot(contains('player?.seek(to: .zero)')),
-          reason:
-              'Looping must not restart by seeking after the item has already '
-              'finished.',
-        );
-      });
-
-      test(
-        '$platform reports requested paused position during texture refresh',
-        () {
-          final source = _appleSourceFile(platform).readAsStringSync();
-
-          expect(
-            source,
-            contains('reportedPositionOverrideMs'),
-            reason:
-                'Paused AVPlayer texture refresh can settle one decoded frame '
-                'behind the requested seek target; native state updates should '
-                'keep reporting the requested timeline position.',
-          );
-          expect(
-            source,
-            contains('if let overrideMs = reportedPositionOverrideMs'),
-          );
-          expect(source, contains('Int64(actualPositionMs) >= overrideMs'));
-          expect(source, contains('reportedPositionOverrideMs = nil'));
-        },
+      expect(
+        source,
+        contains('Thread.isMainThread'),
+        reason:
+            'FlutterEventSink must not be called from AVFoundation/KVO '
+            'callback queues.',
       );
+      expect(source, contains('DispatchQueue.main.async'));
+      expect(source, contains('sendStateUpdateOnMain'));
+    });
 
-      test(
-        '$platform emits freeze diagnostics for stalled native playback',
-        () {
-          final source = _appleSourceFile(platform).readAsStringSync();
+    test('uses native queue-based looping', () {
+      final source = _appleSourceFile().readAsStringSync();
 
-          expect(source, contains('setClipsTimeoutMs'));
-          expect(source, contains('bufferingStallMs'));
-          expect(source, contains('DivineVideoPlayer.Freeze'));
-          expect(source, contains('observeBuffering(for: item)'));
-          expect(source, contains(r'\.isPlaybackBufferEmpty'));
-          expect(source, contains(r'\.isPlaybackLikelyToKeepUp'));
-          expect(source, contains('clearSetClipsTimeout()'));
-          expect(
-            source,
-            contains('clearBufferingWatchdog(resetReported: true)'),
-          );
-        },
+      expect(
+        source,
+        contains('AVQueuePlayer'),
+        reason:
+            'Apple loop playback should use a queue player so the next '
+            'iteration can be prepared before the current one ends.',
       );
-    }
+      expect(
+        source,
+        contains('AVPlayerLooper'),
+        reason:
+            'Manual AVPlayerItemDidPlayToEndTime seek-and-replay loops '
+            'can create an end-of-item playback gap.',
+      );
+      expect(
+        source,
+        isNot(contains('player?.seek(to: .zero)')),
+        reason:
+            'Looping must not restart by seeking after the item has already '
+            'finished.',
+      );
+    });
+
+    test('reports requested paused position during texture refresh', () {
+      final source = _appleSourceFile().readAsStringSync();
+
+      expect(
+        source,
+        contains('reportedPositionOverrideMs'),
+        reason:
+            'Paused AVPlayer texture refresh can settle one decoded frame '
+            'behind the requested seek target; native state updates should '
+            'keep reporting the requested timeline position.',
+      );
+      expect(
+        source,
+        contains('if let overrideMs = reportedPositionOverrideMs'),
+      );
+      expect(source, contains('Int64(actualPositionMs) >= overrideMs'));
+      expect(source, contains('reportedPositionOverrideMs = nil'));
+    });
+
+    test('emits freeze diagnostics for stalled native playback', () {
+      final source = _appleSourceFile().readAsStringSync();
+
+      expect(source, contains('setClipsTimeoutMs'));
+      expect(source, contains('bufferingStallMs'));
+      expect(source, contains('DivineVideoPlayer.Freeze'));
+      expect(source, contains('observeBuffering(for: item)'));
+      expect(source, contains(r'\.isPlaybackBufferEmpty'));
+      expect(source, contains(r'\.isPlaybackLikelyToKeepUp'));
+      expect(source, contains('clearSetClipsTimeout()'));
+      expect(
+        source,
+        contains('clearBufferingWatchdog(resetReported: true)'),
+      );
+    });
   });
 }
 
-File _appleSourceFile(String platform) {
+/// The iOS and macOS players share a single Darwin source tree
+/// (`darwin/divine_video_player/Sources/`), so the threading contract
+/// is asserted once.
+File _appleSourceFile() {
   final packageRelative = File(
-    '$platform/Classes/DivineVideoPlayerInstance.swift',
+    'darwin/divine_video_player/Sources/divine_video_player/'
+    'DivineVideoPlayerInstance.swift',
   );
   if (packageRelative.existsSync()) {
     return packageRelative;
@@ -99,6 +95,7 @@ File _appleSourceFile(String platform) {
 
   return File(
     'packages/divine_video_player/'
-    '$platform/Classes/DivineVideoPlayerInstance.swift',
+    'darwin/divine_video_player/Sources/divine_video_player/'
+    'DivineVideoPlayerInstance.swift',
   );
 }
